fix the example input and output for pset2 problem 5
Big mistake on my part! The example input in pset2 readme.md was using
the wrong key for problem 5. However, the example_input.json file
provided in the repo is correct. This commit updates readme.md to match
the checked in .json files.

Big apologies to students who ran into this!